HRank-master/requirement.txt
absl-py==1.0.0
astroid==2.9.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
Cython==0.29.26
fonttools==4.28.5
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.0
isort==5.10.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
llvmlite==0.37.0
loguru==0.5.3
Markdown==3.3.6
matplotlib==3.5.1
mccabe==0.6.1
networkx==2.6.3
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.1.5
Pillow==9.0.0
platformdirs==2.4.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pylint==2.12.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scipy==1.7.3
Shapely==1.8.0
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
toml==0.10.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.62.3
typed-ast==1.5.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wget==3.2
wrapt==1.13.3
yacs==0.1.8
yapf==0.32.0
zipp==3.7.0
